Add list product screen

+++ guanacos_api/app/models/Product.js
@@ -3,7 +3,7 @@
 module.exports = (sequelize, DataTypes) => {
   const Product = sequelize.define('Product',{
     name: {type: DataTypes.STRING,field:'NAME'},
-    price: {type: DataTypes.DECIMAL,field:'PRICE'},
+    price: {type: DataTypes.DOUBLE,field:'PRICE'},
     description: {type: DataTypes.STRING,field:'DESCRIPTION'},
     image1: {type: DataTypes.STRING,field:'IMAGE1'},
     image2: {type: DataTypes.STRING,field:'IMAGE2'},
